requirements.txt
openai
pandas
streamlit
playwright
python-jose
fastapi
stripe
python-dotenv
requests
uvicorn
'pydantic[email]'